+++ tests/Http/Controllers/CommentsController.php
@@ -3,6 +3,7 @@
 namespace CloudCreativity\LaravelJsonApi\Tests\Http\Controllers;
 
 use CloudCreativity\LaravelJsonApi\Http\Controllers\EloquentController;
+use CloudCreativity\LaravelJsonApi\Tests\JsonApi\Comments\Hydrator;
 use CloudCreativity\LaravelJsonApi\Tests\Models\Comment;
 
 class CommentsController extends EloquentController
@@ -10,9 +11,11 @@ class CommentsController extends EloquentController
 
     /**
      * CommentsController constructor.
+     *
+     * @param Hydrator $hydrator
      */
-    public function __construct()
+    public function __construct(Hydrator $hydrator)
     {
-        parent::__construct(new Comment());
+        parent::__construct(new Comment(), $hydrator);
     }
 }